Refactored recipient´s anticipation limits to getMin and getMax
