+++ src/main/java/me/pagar/model/BulkAnticipationLimits.java
@@ -16,23 +16,4 @@ public class BulkAnticipationLimits {
     public Limit getMinimum() {
         return minimum;
     }
-
-    public class Limit{
-        @Expose(serialize = false)
-        private Integer amount;
-        @Expose(serialize = false)
-        private Integer anticipationFee;
-        @Expose(serialize = false)
-        private Integer fee;
-
-        public Integer getAmount() {
-            return amount;
-        }
-        public Integer getAnticipationFee() {
-            return anticipationFee;
-        }
-        public Integer getFee() {
-            return fee;
-        }
-    }
 }
